+++ PCA9685_proc.h
@@ -42,8 +42,8 @@
 typedef struct PCA9685_t {
 	struct {
 	I2C_slave_t *	slave_inst;		//!< Slave structure
-	uint16_t		Frequency;		//!< Outputs frequency
 	uint32_t		Clock;			//!< Clock frequency
+	float			Frequency;		//!< Outputs frequency
 	} cfg;
 } PCA9685_t;
 
@@ -59,7 +59,6 @@ extern PCA9685_t	PCA9685[I2C_PCA9685_NB];	//!< PCA9685 User structure
 
 /*!\brief Initialization Sequence for PCA9685 peripheral
 ** \weak PCA9685 Init sequence may be user implemented if custom initialization sequence needed
-**
 ** \param[in] pCpnt - Pointer to PCA9685 component
 ** \return FctERR - ErrorCode
 **/
@@ -67,35 +66,22 @@ FctERR NONNULL__ PCA9685_Init_Sequence(PCA9685_t * pCpnt);
 
 
 /*!\brief Sets external clock frequency
-**
 ** \param[in,out] pCpnt - Pointer to PCA9685 component
 ** \param[in] freq - External clock frequency in Hz
 ** \return FctERR - ErrorCode
 **/
-__INLINE FctERR NONNULL_INLINE__ PCA9685_Set_Clock_Frequency(PCA9685_t * pCpnt, const uint32_t freq) {
+__INLINE FctERR NONNULL_INLINE__ PCA9685_Set_Clock_Freq(PCA9685_t * pCpnt, const uint32_t freq) {
 	if (freq > PCA9685_MAX_CLOCK_FREQ)	{ return ERROR_VALUE; }
 	pCpnt->cfg.Clock = freq;
 	return ERROR_OK; }
 
 
-/*!\brief Sets register value for PCA9685 according to desired output frequency
-** \warning if wrong frequency given, value for 500Hz will be returned
-**
+/*!\brief Gets external clock frequency
 ** \param[in] pCpnt - Pointer to PCA9685 component
-** \param[in] freq - Outputs frequency in Hz
-** \return prescaler register value for PCA9685
+** \return Internal clock frequency
 **/
-uint8_t PCA9685_Get_PWM_Prescaler(PCA9685_t * pCpnt, uint16_t freq);
-
-
-/*!\brief Get frequency from prescaler register value of PCA9685
-**
-** \param[in] pCpnt - Pointer to PCA9685 component
-** \param[in] reg - Prescaler register value
-** \return Frequency of PCA9685 outputs
-**/
-__INLINE uint16_t NONNULL__ PCA9685_Get_PWM_Frequency(PCA9685_t * pCpnt, uint8_t reg) {
-	return (uint16_t) (pCpnt->cfg.Clock / (4096 * (reg + 1))); }
+__INLINE uint32_t NONNULL_INLINE__ PCA9685_Get_Clock_Freq(PCA9685_t * pCpnt) {
+	return pCpnt->cfg.Clock; }
 
 
 /****************************************************************/